+++ frontend/src/components/pages/user/LoginPage.tsx
@@ -1,8 +1,9 @@
 import { useState, ChangeEvent,memo, VFC } from "react"
-import { Box, Flex, Heading, Divider, Button, Input, Stack} from "@chakra-ui/react"
+import { Box, Flex, Heading, Divider, Input, Stack} from "@chakra-ui/react"
 
 import { useUserLogIn } from "../../../hooks/useUserLogin";
 import { Link } from "react-router-dom";
+import { BlueButton  } from "../../atms/button/LoginButton";
 
 export const LoginPage: VFC = memo(() => {
   const [inputEmail, setinputEmail] = useState<string>('')
@@ -16,7 +17,7 @@ export const LoginPage: VFC = memo(() => {
   }
 
   return (
-      <Flex align="center" justify="center" height="90vh">
+    <Flex align="center" justify="center" height="90vh">
       <Box bg="white" w="sm" p={4} borderRadius="md" shadow="md">
           <Heading as="h1" size="lg" textAlign="center" py={4}>ログイン</Heading>
           <Divider my={4}></Divider>
@@ -33,20 +34,19 @@ export const LoginPage: VFC = memo(() => {
               value={inputPassword}
               onChange={onChangeInputUserPassword}
             />
-            <Button
-              bg="blue.500"
-              color="white"
-              _hover={{ opacity: 0.7}}
+            <BlueButton
               onClick= {onClickLogin}
-              isLoading = {loading}
-              isDisabled = {inputPassword === '' || inputPassword === ''}
-              >ログイン</Button>
+              loading = {loading}
+              inputEmail={inputEmail}
+              inputPassword={inputPassword}
+              >
+              ログイン
+            </BlueButton>
             <Box  pt={4} align="center">
               <Link to="/signup">新規作成する</Link>
             </Box>
-          </Stack>
+        </Stack>
       </Box>
-      </Flex>
-    
+    </Flex>
   );
 });
